+++ data/src/main/java/com/kitching/data/datasource/impl/OrderDataSourceImpl.kt
@@ -1,6 +1,7 @@
-package com.kitching.data.datasource
+package com.kitching.data.datasource.impl
 
 import com.google.firebase.firestore.FirebaseFirestore
+import com.kitching.data.datasource.OrderDataSource
 import com.kitching.data.dto.OrderCategoryDTO
 import com.kitching.data.dto.OrderDTO
 import com.kitching.data.firebase.COLLECTION_ORDER
@@ -8,7 +9,8 @@ import com.kitching.data.firebase.COLLECTION_ORDER_CATEGORY
 import com.kitching.data.firebase.DOCUMENT_TEAM_ID
 import kotlinx.coroutines.tasks.await
 
-class OrderDataSourceImpl(private val db: FirebaseFirestore = FirebaseFirestore.getInstance()) : OrderDataSource {
+class OrderDataSourceImpl(private val db: FirebaseFirestore = FirebaseFirestore.getInstance()) :
+    OrderDataSource {
 
     override suspend fun getOrderCategories(teamId: String): List<OrderCategoryDTO> {
         return db.collection(COLLECTION_ORDER_CATEGORY)
