+++ data/src/main/java/com/kitching/data/repository/FcmTokenRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.kitching.data.repository
 
 import com.kitching.data.datasource.FcmTokenDataSource
-import com.kitching.data.datasource.FcmTokenDataSourceImpl
+import com.kitching.data.datasource.impl.FcmTokenDataSourceImpl
 import com.kitching.domain.repository.FcmTokenRepository
 import com.kitching.domain.util.AppResult
 import kotlinx.coroutines.flow.Flow
