+++ data/src/main/java/com/kitching/data/repository/LoginRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.kitching.data.repository
 
 import com.kitching.data.datasource.LoginDataSource
-import com.kitching.data.datasource.LoginDataSourceImpl
+import com.kitching.data.datasource.impl.LoginDataSourceImpl
 import com.kitching.domain.entities.User
 import com.kitching.domain.repository.LoginRepository
 import com.kitching.domain.util.AppResult
